% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{get_sent_values}
\alias{get_sent_values}
\title{Assigns Sentiment Values}
\usage{
get_sent_values(char_v, method = "bing")
}
\arguments{
\item{char_v}{A string}

\item{method}{A string indicating which sentiment dictionary to use}
}
\value{
A single numerical value (positive or negative)
based on the assessed sentiment in the string
}
\description{
Assigns sentiment values to words based on preloaded dictionary
}

