% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{get_stanford_sentiment}
\alias{get_stanford_sentiment}
\title{Get Sentiment from the Stanford Tagger}
\usage{
get_stanford_sentiment(text_vector, path_to_stanford_tagger)
}
\arguments{
\item{text_vector}{A vector of strings}

\item{path_to_stanford_tagger}{a local file path indicating 
where the coreNLP package is installed.}
}
\description{
Call the Stanford Sentiment tagger with a
vector of strings.  The Stanford tagger automatically
detects sentence boundaries and treats each sentence as a 
distinct instance to measure. As a result, the vector 
that gets returned will not be the same length as the
input vector.
}
