% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{get_tokens}
\alias{get_tokens}
\title{Word Tokenization}
\usage{
get_tokens(text_of_file, pattern = "\\\\W")
}
\arguments{
\item{text_of_file}{A Text String}

\item{pattern}{A regular expression for token breaking}
}
\value{
A Character Vector of Words
}
\description{
Parses a string into a vector of word tokens.
}

