% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{rescale}
\alias{rescale}
\title{Vector Value Rescaling
 Rescale Transformed values from -1 to 1}
\usage{
rescale(x)
}
\arguments{
\item{x}{A vector of values}
}
\description{
Vector Value Rescaling
 Rescale Transformed values from -1 to 1
}

