% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{rescale2}
\alias{rescale2}
\title{Bi-Directional x and y axis Rescaling}
\usage{
rescale2(v)
}
\arguments{
\item{v}{A vector of values}
}
\description{
Rescale Transformed values from -1 to 1 on the y-axis and scale from zero to 1 on the x-axis
}

