% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{get_nrc_values}
\alias{get_nrc_values}
\title{Summarize NRC values}
\usage{
get_nrc_values(word_vector)
}
\arguments{
\item{word_vector}{A character verctor}
}
\value{
a vector of values for the emotions and valence
detected in the input vector
}
\description{
Access the NRC dictionary to compute emotion types and
valence for a set of words in the input vector
}

