% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{get_percentage_values}
\alias{get_percentage_values}
\title{Chunk a Text and Get Means}
\usage{
get_percentage_values(raw_values, bins = 100)
}
\arguments{
\item{raw_values}{Raw sentiment values}

\item{bins}{The number of bins to split the input vector.
Default is 100 bins.}
}
\value{
A vector of mean values from each chunk
}
\description{
Chunks text into 100 Percentage based segments and calculates means.
}

