% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{get_percentage_values}
\alias{get_percentage_values}
\title{Chunk text into 100 Percentage based segments and calculate means}
\usage{
get_percentage_values(raw_values)
}
\arguments{
\item{raw_values}{Raw sentiment values}
}
\value{
A vector of mean values from each chunk
}
\description{
Chunk text into 100 Percentage based segments and calculate means
}

