% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{get_sentences}
\alias{get_sentences}
\title{Sentence Tokenization
Parses a string into a vector of sentences}
\usage{
get_sentences(text_of_file)
}
\arguments{
\item{text_of_file}{A Text String}
}
\value{
A Character Vector of Sentences
}
\description{
Sentence Tokenization
Parses a string into a vector of sentences
}

