% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{get_sentences}
\alias{get_sentences}
\title{Sentence Tokenization}
\usage{
get_sentences(text_of_file, strip_quotes = TRUE)
}
\arguments{
\item{text_of_file}{A Text String}

\item{strip_quotes}{Logical. Default of TRUE results in 
removal of quote marks from text input prior to sentence 
parsing.}
}
\value{
A Character Vector of Sentences
}
\description{
Parses a string into a vector of sentences.
}

