% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{get_sentiment_dictionary}
\alias{get_sentiment_dictionary}
\title{Sentiment Dictionaries}
\usage{
get_sentiment_dictionary(dictionary = "syuzhet", language = "english")
}
\arguments{
\item{dictionary}{A string indicating which sentiment dictionary to return.  Options include "syuzhet", "bing", "afinn", and "nrc".}

\item{language}{A string indicating the language to choose if using the NRC dictionary and a language other than English}
}
\value{
A \code{\link[base]{data.frame}}
}
\description{
Get the sentiment dictionaries used in \pkg{syuzhet}.
}
\examples{
get_sentiment_dictionary()
get_sentiment_dictionary('bing')
get_sentiment_dictionary('afinn')
get_sentiment_dictionary('nrc', language = "spanish")
}
