% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{get_text_as_string}
\alias{get_text_as_string}
\title{Load Text from File
Load a file as a single text sting.}
\usage{
get_text_as_string(path_to_file)
}
\arguments{
\item{path_to_file}{file path}
}
\description{
Load Text from File
Load a file as a single text sting.
}

