% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{get_text_as_string}
\alias{get_text_as_string}
\title{Load Text from a File}
\usage{
get_text_as_string(path_to_file)
}
\arguments{
\item{path_to_file}{file path}
}
\value{
A character vector of length 1 containing the text of the file in the path_to_file argument.
}
\description{
Loads a file as a single text sting.
}

