% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{simple_plot}
\alias{simple_plot}
\title{Plots simple and rolling shapes overlayed}
\usage{
simple_plot(raw_values, title = "Syuzhet Plot", legend_pos = "top",
  lps = 10, window = 0.1)
}
\arguments{
\item{raw_values}{the raw sentiment values
calculated for each sentence}

\item{title}{for resulting image}

\item{legend_pos}{positon for legend}

\item{lps}{}

\item{window}{}
}
\description{
A Simple function for comparing three smoothers
}

