% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{simple_plot}
\alias{simple_plot}
\title{Plots simple and rolling shapes overlayed}
\usage{
simple_plot(raw_values, title = "Syuzhet Plot", legend_pos = "top")
}
\arguments{
\item{raw_values}{the raw sentiment values
calculated for each sentence}

\item{title}{for image}

\item{legend_pos}{positon for legend}
}
\description{
A Simple function for comparing three smoothers
}

